+++ src/BrokerLib/dcc/DccLiteService.h
@@ -183,7 +183,7 @@ namespace dcclite::broker
 
 			Decoder& Device_CreateDecoder(
 				IDevice_DecoderServices &dev,
-				const std::string &className,
+				std::string_view className,
 				DccAddress address,
 				RName name,
 				const rapidjson::Value &params